create_report/requirements.txt
google-cloud-bigquery
pandas
matplotlib
jupyter
nbconvert
